+++ 01_Nbody/lib/Vector3d.cpp
@@ -2,7 +2,7 @@
 
 // Constructors
 Vector3d::Vector3d() : x(0), y(0), z(0) {}
-Vector3d::Vector3d(double x, double y, double z) : x(x), y(y), z(z) {}
+Vector3d::Vector3d(long double x, long double y, long double z) : x(x), y(y), z(z) {}
 
 // Getters
 double Vector3d::getX() const { return x; }
@@ -10,13 +10,13 @@ double Vector3d::getY() const { return y; }
 double Vector3d::getZ() const { return z; }
 
 // Setters
-void Vector3d::setX(double x) { this->x = x; }
-void Vector3d::setY(double y) { this->y = y; }
-void Vector3d::setZ(double z) { this->z = z; }
+void Vector3d::setX(long double x) { this->x = x; }
+void Vector3d::setY(long double y) { this->y = y; }
+void Vector3d::setZ(long double z) { this->z = z; }
 
 // Member functions
-double Vector3d::magnitude() const { return std::sqrt(x * x + y * y + z * z); }
-Vector3d Vector3d::normalize() const { double mag = magnitude(); return Vector3d(x / mag, y / mag, z / mag); }
+long double Vector3d::magnitude() const { return std::sqrt(x * x + y * y + z * z); }
+Vector3d Vector3d::normalize() const { long double mag = magnitude(); return Vector3d(x / mag, y / mag, z / mag); }
 
 /**
  * @brief Computes the scalar (dot) product of this vector with another vector.
@@ -26,7 +26,7 @@ Vector3d Vector3d::normalize() const { double mag = magnitude(); return Vector3d
  * @param other The vector to compute the scalar product with.
  * @return The scalar product of this vector and the other vector.
  */
-double Vector3d::scalar_product(Vector3d other) const { return (x * other.x + y * other.y + z * other.z); }
+long double Vector3d::scalar_product(Vector3d other) const { return (x * other.x + y * other.y + z * other.z); }
 
 /**
  * @brief Computes the cross product of this vector with another vector.
@@ -41,6 +41,6 @@ Vector3d Vector3d::cross_product(Vector3d other) const { return Vector3d(y * oth
 // Operator overloads
 Vector3d Vector3d::operator+(const Vector3d& other) const { return Vector3d(x + other.x, y + other.y, z + other.z); }
 Vector3d Vector3d::operator-(const Vector3d& other) const { return Vector3d(x - other.x, y - other.y, z - other.z); }
-Vector3d Vector3d::operator*(double scalar) const { return Vector3d(x * scalar, y * scalar, z * scalar); }
-Vector3d Vector3d::operator/(double scalar) const { return Vector3d(x / scalar, y / scalar, z / scalar); }
+Vector3d Vector3d::operator*(long double scalar) const { return Vector3d(x * scalar, y * scalar, z * scalar); }
+Vector3d Vector3d::operator/(long double scalar) const { return Vector3d(x / scalar, y / scalar, z / scalar); }
 Vector3d Vector3d::operator+=(const Vector3d& other) { x += other.x; y += other.y; z += other.z; return *this; }
