+++ 01_Nbody/lib/Velocity_Verlet.hpp
@@ -3,6 +3,6 @@
 #include "Integrator.hpp"
 class Velocity_Verlet : public Integrator {
     public:
-            std::vector<Body> integrate(const std::vector<Body>& bodies, const double maxTimeStep) override;
+            std::vector<Body> integrate(const std::vector<Body>& bodies, const long double maxTimeStep) override;
 };
 #endif // VELOCITY_VERLET_HPP
